back/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
click==8.1.6
colorama==0.4.6
distlib==0.3.6
filelock==3.8.0
Flask==2.3.2
Flask-Cors==4.0.0
flask-restx==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
marshmallow==3.20.1
nodeenv==1.7.0
packaging==23.1
platformdirs==2.5.2
psycopg2==2.9.7
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
referencing==0.30.0
rpds-py==0.9.2
virtualenv==20.16.5
Werkzeug==2.3.6
